% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addExperiments.R
\name{addExperiments}
\alias{addExperiments}
\title{Add Experiments to the Registry}
\usage{
addExperiments(prob.designs = NULL, algo.designs = NULL, repls = 1L,
  combine = "crossprod", reg = getDefaultRegistry())
}
\arguments{
\item{prob.designs}{[named list of \code{\link[data.table]{data.table}} or \code{\link[base]{data.frame}}]\cr
Named list of data frames. The name must match the problem name while the column names correspond to parameters
of the problem.
If \code{NULL}, adds experiments for all defined problems without any hyperparameters set.}

\item{algo.designs}{[named list of \code{\link[data.table]{data.table}} or \code{\link[base]{data.frame}}]\cr
Named list of data frames. The name must match the algorithm name while the column names correspond to parameters
of the algorithm.
If \code{NULL}, adds experiments for all defined algorithms without any hyperparameters set.}

\item{repls}{[\code{integer(1)}]\cr
Number of replications for each distinct experiment.}

\item{combine}{[\code{character(1)}]\cr
How to combine parameters of the single  problem design with a single  algorithm design?
Default is \dQuote{crossprod} which does a \code{\link[data.table]{CJ}} (similar to \code{\link[base]{expand.grid}}
on the problem and algorithm design.
Set to \dQuote{bind} to just \code{\link[base]{cbind}} them where the shorter table is repeated if necessary.}

\item{reg}{[\code{\link{ExperimentRegistry}}]\cr
Registry. If not explicitly passed, uses the last created registry.}
}
\value{
[\code{\link{data.table}}]. Generated job ids are stored in the column \dQuote{job.id}.
  See \code{\link{JoinTables}} for examples on working with job tables.
}
\description{
Adds experiments for running algorithms on problems to the registry and thereby defines batch jobs.
If multiple problem designs or algorithm designs are provided, they are combined via the Cartesian product.
Each row of a single problem design is combined with each row of a single algorithm design in either a
\code{\link[base]{expand.grid}} or \code{\link[base]{cbind}} fashion (depending on parameter \code{combine}).
}
\examples{
reg = makeExperimentRegistry(file.dir = NA, make.default = FALSE)
addProblem(reg = reg, "p1",
  fun = function(job, data, n, mean, sd, ...) rnorm(n, mean = mean, sd = sd))
addAlgorithm(reg = reg, "a1", fun = function(job, data, instance, ...) mean(instance))
prob.designs = list(p1 = expand.grid(n = 100, mean = -3:3, sd = 1:5))
algo.designs = list(a1 = data.table())
addExperiments(reg = reg, prob.designs, algo.designs)
}
\seealso{
Other Experiment: \code{\link{ExperimentRegistry}},
  \code{\link{removeExperiments}},
  \code{\link{summarizeExperiments}}
}

