% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Registry.R
\name{assertRegistry}
\alias{assertRegistry}
\title{assertRegistry}
\usage{
assertRegistry(reg, writeable = FALSE, sync = FALSE, strict = FALSE,
  running.ok = TRUE)
}
\arguments{
\item{reg}{The object asserted to be a \code{Registry}}

\item{writeable}{Optional check if the registry is writeable}

\item{sync}{Whether to sync the registry before writing}

\item{strict}{If \code{FALSE} allow subclasses of \code{Registry}
otherwise.}

\item{running.ok}{If \code{FALSE} throw an error if jobs associated
with the registry are currently running.}
}
\value{
\code{TRUE} invisibily
}
\description{
Assert that a given object is a \code{batchtools} registry.
Additionally can sync the registry, check if it is writeable,
check if it is running.
}
\details{
If any check fails, throws an error indicting
the reason for the failure.
}
