% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export.R
\name{batchExport}
\alias{batchExport}
\title{Export Objects to the Slaves}
\usage{
batchExport(x = list(), reg = getDefaultRegistry())
}
\arguments{
\item{x}{[\code{list}]\cr
Named list of objects to export.
Set to \code{NULL} to unexport the object.}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explicitly passed, uses the last created registry.}
}
\value{
[\code{data.table}] with name and uri to the exported objects.
}
\description{
Objects are saved in subdirectory \dQuote{exports} of the
\dQuote{file.dir} of \code{reg}.
They are automatically loaded and put into the global environment
each time the registry is loaded on the master as well as on the slaves.
}
\examples{
tmp = makeRegistry(file.dir = NA, make.default = FALSE)

# list exports
exports = batchExport(reg = tmp)
print(exports)

# add job depending on exports
batchMap(function(x) x^2 + y + z, x = 1:3, reg = tmp)
exports = batchExport(list(y = 99, z = 1), reg = tmp)
print(exports)

submitJobs(reg = tmp)
waitForJobs(reg = tmp)
loadResult(1, reg = tmp)

# Un-export z
exports = batchExport(list(z = NULL), reg = tmp)
print(exports)
}

