% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchmark.r
\name{batchmark}
\alias{batchmark}
\title{Run Machine Learning Benchmarks as Experiments}
\usage{
batchmark(learners, tasks, resamplings, measures = NULL, repls = 1L,
  save.models = FALSE, reg = getDefaultRegistry())
}
\arguments{
\item{learners}{[list of \link[mlr]{Learner}]\cr
See \link[mlr]{Learner}].}

\item{tasks}{[list of \link[mlr]{Task}]\cr
See \link[mlr]{Task}].}

\item{resamplings}{[list of \link[mlr]{ResampleDesc} or \link[mlr]{ResampleInstance}]\cr
See \link[mlr]{ResampleDesc} or \link[mlr]{ResampleInstance}.}

\item{measures}{[list of \link[mlr]{Measure}]\cr
See \link[mlr]{Measure}.}

\item{repls}{[code{integer(1)}]\cr
Number of replications for each experiment.}

\item{save.models}{[code{logical(1)}]\cr
Store (potentially large) models?}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explicitly passed, uses the last created registry.}
}
\value{
[\code{data.table}]. Generated job ids are stored in the column \dQuote{job.id}.
}
\description{
This function is a very parallel version of \code{\link[mlr]{benchmark}}.
Experiments are created in the provided registry for each combination of
learners, tasks and resamplings.
}

