% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{batchtools-deprecated}
\alias{batchtools-deprecated}
\title{Deprecated function in the batchtools package}
\description{
The following functions have been deprecated:
\tabular{rl}{
  \code{chunkIds} \tab deprecated in favor of \code{\link{chunk}}, \code{\link{lpt}} and \code{\link{binpack}}\cr
}
}
