% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{package}
\name{batchtools-package}
\alias{batchtools}
\alias{batchtools-package}
\title{Tools for computation on batch systems}
\description{
For bug reports and feature requests please use the tracker:
\url{https://github.com/mllg/batchtools}.
}
\section{Package options}{

\describe{
  \item{\code{batchtools.temp.dir}}{
    Temporary directory to use in \code{\link{makeTempRegistry}}.
    On some cluster systems you might want to set this to a directory which is shared between nodes.
    Default is the system temp directory as returned by \code{\link[base]{tempdir}}.
  }
  \item{\code{batchtools.verbose}}{
    Verbosity. Set to \code{FALSE} to suppress info messages and progress bars.
  }
}
}

