% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterFunctions.R
\name{cfBrewTemplate}
\alias{cfBrewTemplate}
\title{Cluster Functions Helper to Write Job Description Files}
\usage{
cfBrewTemplate(reg, text, jc)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explicitly passed, uses the last created registry.}

\item{text}{[\code{character(1)}]\cr
String ready to be brewed. See \code{\link{cfReadBrewTemplate}} to read a template from the file system.}

\item{jc}{[\code{\link{JobCollection})}]\cr
JobCollection holding all essential information. Will be used as environment to look
up variables.}
}
\value{
[\code{character(1)}]. File path to resulting template file.
}
\description{
This function is only intended for use in your own cluster functions implementation.

Calls brew silently on your template, any error will lead to an exception. If debug mode is
enabled (via option \dQuote{batchtools.debug}), the file is stored at the same place as the corresponding
R script in the \dQuote{jobs}-subdir of your files directory, otherwise in the temp dir via
\code{\link{tempfile}}.
}
\seealso{
Other ClusterFunctionsHelper: \code{\link{cfHandleUnknownSubmitError}},
  \code{\link{cfKillJob}},
  \code{\link{cfReadBrewTemplate}},
  \code{\link{makeSubmitJobResult}}
}

