% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterFunctions.R
\name{cfReadBrewTemplate}
\alias{cfReadBrewTemplate}
\title{Cluster Functions Helper to Parse a Brew Template}
\usage{
cfReadBrewTemplate(template = NULL, text = NULL,
  comment.string = NA_character_)
}
\arguments{
\item{template}{[\code{character(1)}]\cr
Path to a brew template file that can be used for the job description.
Mutually exclusive with argument \code{text}.}

\item{text}{[\code{character(1)}]\cr
Single string holding the template to be used.
Mutually exclusive with argument \code{template}.}

\item{comment.string}{[\code{character(1)}]\cr
Ignore lines starting with this string.}
}
\value{
[\code{character}].
}
\description{
This function is only intended for use in your own cluster functions implementation.
Simply reads your template and returns it as a character vector.
}
\seealso{
Other ClusterFunctionsHelper: \code{\link{cfBrewTemplate}},
  \code{\link{cfHandleUnknownSubmitError}},
  \code{\link{cfKillJob}},
  \code{\link{makeSubmitJobResult}}
}

