% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chunkIds.r
\name{chunkIds}
\alias{chunkIds}
\title{Chunk Jobs for Sequential Execution}
\usage{
chunkIds(ids = NULL, n.chunks = NULL, chunk.size = NULL,
  group.by = character(0L), reg = getDefaultRegistry())
}
\arguments{
\item{ids}{[\code{\link[base]{data.frame}} or \code{integer}]\cr
A \code{\link[base]{data.frame}} (or \code{\link[data.table]{data.table}})
with a column named \dQuote{job.id}.
This is usually the result of a \code{\link{findJobs}} operation, but you may also
pass any other table here, e.g. the return of \code{\link{getJobInfo}}.
Alternatively, you may also pass a vector of integerish job ids.
If not set, defaults to all jobs.}

\item{n.chunks}{[\code{integer(1)}]\cr
Requested number of chunks.
If more chunks than elements in \code{ids} are requested, empty chunks are
dropped. Cannot be used in combination with \code{chunks.size}.}

\item{chunk.size}{[\code{integer(1)}]\cr
Requested number of elements in each chunk.
Cannot be used in combination with \code{n.chunks}.
If \code{ids} cannot be chunked evenly, some chunks will have less elements.}

\item{group.by}{[\code{character(0)}]\cr
If \code{ids} is a \code{\link{data.frame}} with additional columns (besides \dQuote{job.id}),
then the chunking is done in subgroups defined by the columns \code{group.by}.
Usually not needed for a regular \code{\link{Registry}} as the jobs are likely homogeneous.
For an \code{\link{ExperimentRegistry}} on the other hand, you can use \code{group.by}
to first partition the jobs into homogeneous groups and then chunk them (c.f. example).}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explictitly passed, uses the last created registry.}
}
\value{
[\code{\link[data.table]{data.table}}]. Table \code{ids} with additional column
  \dQuote{job.id}.
}
\description{
Groups jobs into chunks which will be executed sequentially on the nodes.
Useful if you have a large number of fast jobs or just to combine multiple
jobs to reach a desired runtime.

Chunks are submitted via \code{\link{submitJobs}} by simply providing passing
job ids as a table with an extra column \dQuote{chunk}.
All jobs with the same chunk number will be grouped together.
}
\examples{
# chunking for Registry
reg = makeTempRegistry(make.default = FALSE)
ids = batchMap(identity, 1:25, reg = reg)
ids = chunkIds(ids, chunk.size = 10, reg = reg)
print(ids)
print(table(ids$chunk))

# chunking for ExperimentRegistry
reg = makeTempExperimentRegistry(make.default = FALSE)
prob = addProblem(reg = reg, "prob1", data = iris, fun = function(data) nrow(data))
prob = addProblem(reg = reg, "prob2", data = Titanic, fun = function(data) nrow(data))
algo = addAlgorithm(reg = reg, "algo", fun = function(problem, i, ...) problem)
prob.designs = list(prob1 = data.table(), prob2 = data.table(x = 1:2))
algo.designs = list(algo = data.table(i = 1:5))
addExperiments(prob.designs, algo.designs, repls = 10, reg = reg)

# -> group into chunks of 5 jobs, but do not mix problems
ids = getJobInfo(reg = reg)[, .(job.id, problem, algorithm)]
ids = chunkIds(ids, chunk.size = 5, group.by = "problem", reg = reg)
print(ids)
dcast(ids, chunk ~ problem)
}

