% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearRegistry.R
\name{clearRegistry}
\alias{clearRegistry}
\title{Remove All Jobs}
\usage{
clearRegistry(reg = getDefaultRegistry())
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explicitly passed, uses the default registry (see \code{\link{setDefaultRegistry}}).}
}
\description{
Removes all jobs from a registry and calls \code{\link{sweepRegistry}}.
}
\seealso{
Other Registry: 
\code{\link{getDefaultRegistry}()},
\code{\link{loadRegistry}()},
\code{\link{makeRegistry}()},
\code{\link{removeRegistry}()},
\code{\link{saveRegistry}()},
\code{\link{sweepRegistry}()},
\code{\link{syncRegistry}()}
}
\concept{Registry}
