% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doJobCollection.R
\name{doJobCollection}
\alias{doJobCollection}
\title{Execute Jobs of a JobCollection}
\usage{
doJobCollection(jc, con = stdout())
}
\arguments{
\item{jc}{[\code{\link{JobCollection}}]\cr
Either an object of class \dQuote{JobCollection} as returned by
\code{\link{makeJobCollection}} or a string point to file containing a
\dQuote{JobCollection} (saved with \code{\link[base]{saveRDS}}).}

\item{con}{[\code{\link[base]{connection}} | \code{character(1)}]\cr
A connection for the output. Defaults to \code{\link[base]{stdout}}.
Alternatively the name of a file to write to.}
}
\value{
[\code{character(1)}]: Hash of the \code{\link{JobCollection}} executed.
}
\description{
Executes every job in a \code{\link{JobCollection}}.
This function is intended to be called on the slave.
}

