% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doJobs.r
\name{doJobs}
\alias{doJobs}
\title{Execute jobs}
\usage{
doJobs(jd, con = stdout())
}
\arguments{
\item{jd}{[\code{\link{JobDescription}}]\cr
Either an object of class \dQuote{JobDescription} as returned by \code{\link{makeJobDescription}} or a
string point to file containing a \dQuote{JobDescription} (saved with \code{\link[base]{saveRDS}}).}

\item{con}{[\code{\link[base]{connection}}]\cr
A connection to redirect the output to.}
}
\value{
[\code{data.table}]. Data table with updates on the computational status, i.e.
  a table with the columns \dQuote{job.id}, \dQuote{started} (unix time stamp of job start),
  \dQuote{done} (unix time stamp of job termination), \dQuote{error} (error message as string)
  and \code{memory} (memory usage as double).
}
\description{
Execute jobs
}

