% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execJob.r
\name{execJob}
\alias{execJob}
\title{Execute a single job}
\usage{
execJob(job)
}
\arguments{
\item{job}{[\code{\link{Job}}]\cr
Job to execute.}
}
\value{
[any]. Result of the job.
}
\description{
Executes a single job (as created by \code{\link{makeJob}}) and returns
its result. Also works for Experiments.
}

