% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findConfFile.R
\name{findConfFile}
\alias{findConfFile}
\title{Search for a Configuration File in Predefined Locations}
\usage{
findConfFile(name)
}
\arguments{
\item{name}{[\code{character(1)}]\cr
File name of the config file.}
}
\value{
[\code{character}]: location of first hit, empty character vector otherwise.
}
\description{
Searches for a valid configuration file in the following locations and returns the first hit:
\enumerate{
  \item In the current working directory: \code{[getwd()]/[name]}
  \item In the home directory (prefixed with a dot): \code{~/.[name]}
}
}
\examples{
# returns the batchtools configuration file
# (from current working dir or home dir)
findConfFile("batchtools.conf.R")
}

