% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefaultRegistry.R
\name{getDefaultRegistry}
\alias{getDefaultRegistry}
\alias{setDefaultRegistry}
\title{Get and Set the Default Registry}
\usage{
getDefaultRegistry()

setDefaultRegistry(reg)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explicitly passed, uses the default registry (see \code{\link{setDefaultRegistry}}).}
}
\description{
\code{getDefaultRegistry} returns the registry currently set as default (or
stops with an exception if none is set). \code{setDefaultRegistry} sets
a registry as default.
}
\seealso{
Other Registry: 
\code{\link{clearRegistry}()},
\code{\link{loadRegistry}()},
\code{\link{makeRegistry}()},
\code{\link{removeRegistry}()},
\code{\link{saveRegistry}()},
\code{\link{sweepRegistry}()},
\code{\link{syncRegistry}()}
}
\concept{Registry}
