% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getErrorMessages.R
\name{getErrorMessages}
\alias{getErrorMessages}
\title{Retrieve Error Messages}
\usage{
getErrorMessages(ids = NULL, missing.as.error = FALSE,
  reg = getDefaultRegistry())
}
\arguments{
\item{ids}{[\code{\link[base]{data.frame}} or \code{integer}]\cr
A \code{\link[base]{data.frame}} (or \code{\link[data.table]{data.table}})
with a column named \dQuote{job.id}.
Alternatively, you may also pass a vector of integerish job ids.
If not set, defaults to the return value of \code{\link{findErrors}}.}

\item{missing.as.error}{[\code{logical(1)}]\cr
Treat missing results as errors? If \code{TRUE}, an error message is imputed for jobs
which are not terminated. Default is \code{FALSE}}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explicitly passed, uses the last created registry.}
}
\value{
[\code{\link{data.table}}]. Table with columns \dQuote{job.id}, \dQuote{terminated} (logical),
  \dQuote{error} (logical) and \dQuote{message} (string).
  See \code{\link{JoinTables}} for examples on working with job tables.
}
\description{
Extracts error messages from the internal data base and returns them in a compact table.
For further investigation of logs, use \code{\link{showLog}}.
}
\examples{
tmp = makeRegistry(file.dir = NA, make.default = FALSE)
fun = function(i) if (i == 3) stop(i) else i
ids = batchMap(fun, i = 1:5, reg = tmp)
submitJobs(1:4, reg = tmp)
waitForJobs(1:4, reg = tmp)
getErrorMessages(ids, reg = tmp)
getErrorMessages(ids, missing.as.error = TRUE, reg = tmp)
}
\seealso{
Other debug: \code{\link{grepLogs}},
  \code{\link{killJobs}}, \code{\link{resetJobs}},
  \code{\link{showLog}}, \code{\link{testJob}}
}

