% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatus.R
\name{getStatus}
\alias{getStatus}
\title{Summarize the Computational Status}
\usage{
getStatus(ids = NULL, reg = getDefaultRegistry())
}
\arguments{
\item{ids}{[\code{\link[base]{data.frame}} or \code{integer}]\cr
A \code{\link[base]{data.frame}} (or \code{\link[data.table]{data.table}})
with a column named \dQuote{job.id}.
Alternatively, you may also pass a vector of integerish job ids.
If not set, defaults to all jobs.}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explicitly passed, uses the last created registry.}
}
\value{
[\code{\link[data.table]{data.table}}] with class \dQuote{Status} to
  trigger a nice print function.
}
\description{
This function gives an encompassing overview over the computational status
on your system.
}

