% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logs.R
\name{grepLogs}
\alias{grepLogs}
\title{Grep Log Files for a Pattern}
\usage{
grepLogs(ids = NULL, pattern, reg = getDefaultRegistry())
}
\arguments{
\item{ids}{[\code{\link[base]{data.frame}} or \code{integer}]\cr
A \code{\link[base]{data.frame}} (or \code{\link[data.table]{data.table}})
with a column named \dQuote{job.id}.
Alternatively, you may also pass a vector of integerish job ids.
If not set, defaults to the return value of \code{\link{findStarted}}.}

\item{pattern}{[\code{character(1L)}]\cr
Regular expression.}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explicitly passed, uses the last created registry.}
}
\value{
[\code{data.table}]. Matching job ids are stored in the column \dQuote{job.id}.
}
\description{
Greps through log files and reports jobs with matches.
}
\seealso{
Other debug: \code{\link{killJobs}}, \code{\link{showLog}},
  \code{\link{testJob}}
}

