% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobCollection.R
\name{makeJobCollection}
\alias{JobCollection}
\alias{makeJobCollection}
\title{JobCollection Constructor}
\usage{
makeJobCollection(ids = NULL, resources = list(),
  reg = getDefaultRegistry())
}
\arguments{
\item{ids}{[\code{\link[base]{data.frame}} or \code{integer}]\cr
A \code{\link[base]{data.frame}} (or \code{\link[data.table]{data.table}})
with a column named \dQuote{job.id}.
Alternatively, you may also pass a vector of integerish job ids.
If not set, defaults to all jobs.}

\item{resources}{[\code{list}]\cr
Named list of resources. Default is \code{list()}.}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explicitly passed, uses the last created registry.}
}
\value{
[\code{JobCollection}].
}
\description{
\code{makeJobCollection} takes multiple job ids and creates an object of class \dQuote{JobCollection} which holds all
necessary information for the calculation with \code{\link{doJobCollection}}. It is implemented as an environment
with the following variables:
\describe{
 \item{file.dir}{\code{file.dir} of the \link{Registry}.}
 \item{work.dir:}{\code{work.dir} of the \link{Registry}.}
 \item{job.hash}{Unique identifier of the job. Used to create names on the file system.}
 \item{log.file}{Location of the designated log file for this job.}
 \item{packages}{Vector of packages to load on the slaves, see \link{Registry}.}
 \item{resources:}{Named list of of specified computational resources.}
 \item{uri}{Location of the job description file (saved with \code{link[base]{saveRDS}} on the file system.}
 \item{defs}{\code{\link[data.table]{data.table}} holding individual job information. See examples.}
}
If your \link{ClusterFunctions} uses a template, \code{\link[brew]{brew}} will be executed in the environment of such
a collection. Thus all variables available inside the job can be used in the template.
}
\examples{
tmp = makeRegistry(file.dir = NA, make.default = FALSE)
batchMap(identity, 1:5, reg = tmp)
jc = makeJobCollection(1:3, reg = tmp)
ls(jc)
jc$defs
jc$defs$pars
}
\seealso{
Other JobCollection: \code{\link{doJobCollection}}
}

