% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/killJobs.r
\name{killJobs}
\alias{killJobs}
\title{Kill Jobs}
\usage{
killJobs(ids = NULL, reg = getDefaultRegistry())
}
\arguments{
\item{ids}{[\code{\link[base]{data.frame}} or \code{integer}]\cr
A \code{\link[base]{data.frame}} (or \code{\link[data.table]{data.table}})
with a column named \dQuote{job.id}.
Alternatively, you may also pass a vector of integerish job ids.
If not set, defaults to the return value of \code{\link{findOnSystem}}.}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explicitly passed, uses the last created registry.}
}
\value{
[\code{data.table}] with columns \dQuote{job.id}, the corresponding \dQuote{batch.id} and
  the flag \dQuote{killed} to indicate success.
}
\description{
Kill jobs which have already been submitted to the batch system.
If a job is killed its internal state is reset as if it had not been submitted at all.

In case of an error when killing, the function tries - after a short sleep - to kill the remaining
batch jobs again. If this fails again for some jobs, the function gives up. Only jobs that could be
killed are reset in the DB.
}
\seealso{
Other debug: \code{\link{grepLogs}}, \code{\link{showLog}},
  \code{\link{testJob}}
}

