% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterFunctionsDockerQueue.R
\name{makeClusterFunctionsDockerQueue}
\alias{makeClusterFunctionsDockerQueue}
\title{ClusterFunctions for DockerQueue

Customized cluster functions for the isolated application running on the SFB876 cluster.}
\usage{
makeClusterFunctionsDockerQueue(image, docker.args = character(0L),
  image.args = character(0L), scheduler.latency = 1, fs.latency = 65,
  docker.scheduler.url = "https://s876cnsm:2350/v1.30",
  curl.args = character(0L))
}
\arguments{
\item{image}{[\code{character(1)}]\cr
Name of the docker image to run.}

\item{docker.args}{[\code{character}]\cr
Additional arguments passed to \dQuote{docker} *before* the command (\dQuote{run}, \dQuote{ps} or \dQuote{kill}) to execute (e.g., the docker host).}

\item{image.args}{[\code{character}]\cr
Additional arguments passed to \dQuote{docker run} (e.g., to define mounts or environment variables).}

\item{scheduler.latency}{[\code{numeric(1)}]\cr
Time to sleep after important interactions with the scheduler to ensure a sane state.
Currently only triggered after calling \code{\link{submitJobs}}.}

\item{fs.latency}{[\code{numeric(1)}]\cr
Expected maximum latency of the file system, in seconds.
Set to a positive number for network file systems like NFS which enables more robust (but also more expensive) mechanisms to
access files and directories.
Usually safe to set to \code{NA} to disable the heuristic, e.g. if you are working on a local file system.}

\item{docker.scheduler.url}{[\code{character}]\cr
URL of the docker scheduler API.}
}
\value{
[\code{\link{ClusterFunctions}}].
}
\seealso{
Other ClusterFunctions: \code{\link{makeClusterFunctionsDocker}},
  \code{\link{makeClusterFunctionsInteractive}},
  \code{\link{makeClusterFunctionsLSF}},
  \code{\link{makeClusterFunctionsMulticore}},
  \code{\link{makeClusterFunctionsOpenLava}},
  \code{\link{makeClusterFunctionsSGE}},
  \code{\link{makeClusterFunctionsSSH}},
  \code{\link{makeClusterFunctionsSlurm}},
  \code{\link{makeClusterFunctionsSocket}},
  \code{\link{makeClusterFunctionsTORQUE}},
  \code{\link{makeClusterFunctions}}
}
