% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterFunctionsHyperQueue.R
\name{makeClusterFunctionsHyperQueue}
\alias{makeClusterFunctionsHyperQueue}
\title{ClusterFunctions for HyperQueue}
\usage{
makeClusterFunctionsHyperQueue(scheduler.latency = 1, fs.latency = 65)
}
\arguments{
\item{scheduler.latency}{[\code{numeric(1)}]\cr
Time to sleep after important interactions with the scheduler to ensure a sane state.
Currently only triggered after calling \code{\link{submitJobs}}.}

\item{fs.latency}{[\code{numeric(1)}]\cr
Expected maximum latency of the file system, in seconds.
Set to a positive number for network file systems like NFS which enables more robust (but also more expensive) mechanisms to
access files and directories.
Usually safe to set to \code{0} to disable the heuristic, e.g. if you are working on a local file system.}
}
\value{
[ClusterFunctions].
}
\description{
Cluster functions for HyperQueue (\url{https://it4innovations.github.io/hyperqueue/stable/}).

Jobs are submitted via the HyperQueue CLI using \code{hq submit} and executed by  calling \code{Rscript -e "batchtools::doJobCollection(...)"}.
The job name is set to the job hash and logs are handled internally by batchtools.
Listing jobs uses \code{hq job list} and cancelling jobs uses \code{hq job cancel}.
A running HyperQueue server and workers are required.
}
\seealso{
Other ClusterFunctions: 
\code{\link{makeClusterFunctions}()},
\code{\link{makeClusterFunctionsDocker}()},
\code{\link{makeClusterFunctionsInteractive}()},
\code{\link{makeClusterFunctionsLSF}()},
\code{\link{makeClusterFunctionsMulticore}()},
\code{\link{makeClusterFunctionsOpenLava}()},
\code{\link{makeClusterFunctionsSGE}()},
\code{\link{makeClusterFunctionsSSH}()},
\code{\link{makeClusterFunctionsSlurm}()},
\code{\link{makeClusterFunctionsSocket}()},
\code{\link{makeClusterFunctionsTORQUE}()}
}
\concept{ClusterFunctions}
