% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterFunctionsOpenLava.R
\name{makeClusterFunctionsOpenLava}
\alias{makeClusterFunctionsOpenLava}
\title{ClusterFunctions for OpenLava}
\usage{
makeClusterFunctionsOpenLava(template)
}
\arguments{
\item{template}{[\code{character(1)}]\cr
Path to a brew template file that is used for the job description.}
}
\value{
[\code{\link{ClusterFunctions}}].
}
\description{
Job files are created based on the brew template \code{template}. This
file is processed with brew and then submitted to the queue using the
\code{bsub} command. Jobs are killed using the \code{bkill} command and the
list of running jobs is retrieved using \code{bjobs -u $USER -w}. The user
must have the appropriate privileges to submit, delete and list jobs on the
cluster (this is usually the case).

The template file can access all arguments passed to the \code{submitJob}
function, see here \code{\link{ClusterFunctions}}. It is the template file's
job to choose a queue for the job and handle the desired resource
allocations.
}
\seealso{
Other ClusterFunctions: \code{\link{makeClusterFunctionsDocker}},
  \code{\link{makeClusterFunctionsLSF}},
  \code{\link{makeClusterFunctionsSGE}},
  \code{\link{makeClusterFunctionsSLURM}},
  \code{\link{makeClusterFunctionsTorque}},
  \code{\link{makeClusterFunctions}}
}

