% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterFunctionsSLURM.R
\name{makeClusterFunctionsSLURM}
\alias{makeClusterFunctionsSLURM}
\title{ClusterFunctions for SLURM Systems}
\usage{
makeClusterFunctionsSLURM(template = NULL, text = NULL, clusters = NULL)
}
\arguments{
\item{template}{[\code{character(1)}]\cr
Path to a brew template file that can be used for the job description.
Mutually exclusive with argument \code{text}.}

\item{text}{[\code{character(1)}]\cr
Single string holding the template to be used.
Mutually exclusive with argument \code{template}.}

\item{clusters}{[\code{character(1)}]\cr
If multiple clusters are managed by one SLURM system, the name of one cluster has to be specified.
If only one cluster is present, this argument may be omitted.
Note that you should not select the cluster in your template file via \code{#SBATCH --clusters}.}
}
\value{
[\code{\link{ClusterFunctions}}].
}
\description{
Job files are created based on the brew template \code{template.file}. This
file is processed with brew and then submitted to the queue using the
\code{sbatch} command. Jobs are killed using the \code{scancel} command and
the list of running jobs is retrieved using \code{squeue}. The user must
have the appropriate privileges to submit, delete and list jobs on the
cluster (this is usually the case).

The template file can access all arguments passed to the \code{submitJob}
function, see here \code{\link{ClusterFunctions}}. It is the template file's
job handle the desired resource allocations. The cluster selection however is
done via the argument \code{clusters}.
}
\seealso{
Other ClusterFunctions: \code{\link{makeClusterFunctionsDocker}},
  \code{\link{makeClusterFunctionsLSF}},
  \code{\link{makeClusterFunctionsOpenLava}},
  \code{\link{makeClusterFunctionsSGE}},
  \code{\link{makeClusterFunctionsTorque}},
  \code{\link{makeClusterFunctions}},
  \code{\link{runOSCommand}}
}

