% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterFunctionsSocket.R
\name{makeClusterFunctionsSocket}
\alias{makeClusterFunctionsSocket}
\title{ClusterFunctions for Parallel Socket Execution}
\usage{
makeClusterFunctionsSocket(ncpus = NA_integer_)
}
\arguments{
\item{ncpus}{[\code{integer(1)}]\cr
Number of VPUs of worker.
Default is to use all cores. The total number of cores "available" is given by option \code{mc.cores}
and defaults to the heuristic implemented in \code{\link[parallel]{detectCores}}.}
}
\value{
[\code{\link{ClusterFunctions}}].
}
\description{
Jobs are spawned asynchronously using the package \pkg{snow}.
}
\seealso{
Other clusterFunctions: \code{\link{makeClusterFunctionsInteractive}},
  \code{\link{makeClusterFunctionsMulticore}},
  \code{\link{makeClusterFunctionsSSH}}
}

