% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterFunctionsSSH.r
\name{makeClusterFunctionsSSH}
\alias{makeClusterFunctionsSSH}
\title{ClusterFunctions for Remote SSH Execution}
\usage{
makeClusterFunctionsSSH(workers)
}
\arguments{
\item{workers}{[\code{list} of \code{\link{Worker}}]\cr
List of Workers as constructed with \code{\link{Worker}}.}
}
\value{
[\code{\link{ClusterFunctions}}].
}
\description{
Jobs are spawned by starting multiple R sessions via SSH.
}
\note{
If you use a custom \dQuote{.ssh/config} file, make sure your
ProxyCommand passes \sQuote{-q} to ssh, otherwise each output will
end with the message \dQuote{Killed by signal 1} and this will break
the communication with the nodes.
}
\seealso{
Other clusterFunctions: \code{\link{makeClusterFunctionsInteractive}},
  \code{\link{makeClusterFunctionsMulticore}}
}

