% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterFunctionsTorque.R
\name{makeClusterFunctionsTorque}
\alias{makeClusterFunctionsTorque}
\title{ClusterFunctions for OpenPBS/Torque Systems}
\usage{
makeClusterFunctionsTorque(template = findTemplateFile("torque"))
}
\arguments{
\item{template}{[\code{character(1)}]\cr
Path to a brew template file that can be used for the job description.
Alternatively, the template as a single string (including at least one newline \dQuote{\\n}).
Defaults to a heuristic which looks for a template file in the following locations and picks
the first one found:
(1) \dQuote{batchtools.torque.tmpl} in the current working directory,
(2) \dQuote{.batchtools.torque.tmpl} in the home directory and finally
(3) \dQuote{torque.default.tmpl} in the package installation directory
in the subfolder \dQuote{templates}.}
}
\value{
[\code{\link{ClusterFunctions}}].
}
\description{
Job files are created based on the brew template \code{template.file}. This file is processed
with brew and then submitted to the queue using the \code{qsub} command. Jobs are killed using
the \code{qdel} command and the list of running jobs is retrieved using \code{qselect}. The user
must have the appropriate privileges to submit, delete and list jobs on the cluster (this is
usually the case).

The template file can access all arguments passed to the \code{submitJob} function, see here
\code{\link{ClusterFunctions}}. It is the template file's job to choose a queue for the job and
handle the desired resource allocations.
}
\seealso{
Other ClusterFunctions: \code{\link{makeClusterFunctionsDocker}},
  \code{\link{makeClusterFunctionsLSF}},
  \code{\link{makeClusterFunctionsMulticore}},
  \code{\link{makeClusterFunctionsOpenLava}},
  \code{\link{makeClusterFunctionsSGE}},
  \code{\link{makeClusterFunctionsSlurm}},
  \code{\link{makeClusterFunctionsSocket}},
  \code{\link{makeClusterFunctions}},
  \code{\link{runOSCommand}}
}

