% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Job.r
\name{makeJob}
\alias{Job}
\alias{makeJob}
\title{Computational Jobs}
\usage{
makeJob(id, reg = getDefaultRegistry())
}
\arguments{
\item{id}{[\code{integer(1)} or \code{data.table}]\cr
Single integer to specify the job or a \code{data.table} with column \code{job.id}
and exactly one row.}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explictitly passed, uses the last created registry.}
}
\value{
[\code{Job}].
}
\description{
\code{makeJob} takes a single job id and creates an object of class
\dQuote{Job}. This job is passed to reduce functions like
\code{\link{reduceResults}} or to \code{\link{Algorithm}} for an
\code{\link{ExperimentRegistry}}. calculation with \code{\link{doJobs}}. It
is implemented as a list holding the following variables:
\describe{
 \item{job.id}{Job ID as integer.}
 \item{pars}{Job parameters. For \code{\link{ExperimentRegistry}}, this is a list with named elements \dQuote{prob.name}, \dQuote{prob.pars}, \dQuote{algo.name} and \dQuote{algo.pars}.}
 \item{seed}{Seed which is set via \code{\link{doJobs}}.}
 \item{resources}{Computational resources which were set for this job.}
}
A regular \code{\link{Registry}} also has the items \dQuote{fun} and an
\code{\link{ExperimentRegistry}} has the slots \dQuote{problem} and
\dQuote{algorithm} (see \code{\link{Problem}}).
}
\examples{
reg = makeTempRegistry(make.default = FALSE)
batchMap(identity, 1:5, reg = reg)
job = makeJob(1, reg = reg)
names(job)
}

