% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results.r
\name{reduceResults}
\alias{reduceResults}
\title{Reduce Results}
\usage{
reduceResults(fun, ids = NULL, init, ..., reg = getDefaultRegistry())
}
\arguments{
\item{fun}{[\code{function}]\cr
A function to reduce the results. The result of previous iterations (or
the \code{init}) will be passed as first argument, the result of of the
i-th iteration as second. See \code{\link[base]{Reduce}} for some
examples. If the function has the formal argument \dQuote{job}, the job
description is passed.}

\item{ids}{[\code{\link[base]{data.frame}} or \code{integer}]\cr
A \code{\link[base]{data.frame}} (or \code{\link[data.table]{data.table}})
with a column named \dQuote{job.id}.
This is usually the result of a \code{\link{findJobs}} operation, but you may also
pass any other table here, e.g. the return of \code{\link{getJobInfo}}.
Alternatively, you may also pass a vector of integerish job ids.
If not set, defaults to result of \code{\link{findDone}}.}

\item{init}{[\code{ANY}]\cr
Initial element, as used in \code{\link[base]{Reduce}}.
Default is the first result.}

\item{...}{[\code{ANY}]\cr
Additional arguments passed to to function \code{fun}.}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explictitly passed, uses the last created registry.}
}
\value{
Aggregated results, return type depends on function. If \code{ids}
is empty, \code{reduceResults} returns \code{init} (if available) or
\code{NULL} otherwise.
}
\description{
A version of \code{\link[base]{Reduce}} for \code{\link{Registry}} objects.
}
\seealso{
Other Results: \code{\link{loadResult}},
  \code{\link{reduceResultsList}}
}

