% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results.r
\name{reduceResultsList}
\alias{reduceResultsDataTable}
\alias{reduceResultsList}
\title{Aggregate Results}
\usage{
reduceResultsList(ids = NULL, fun = NULL, ..., reg = getDefaultRegistry())

reduceResultsDataTable(ids = NULL, fun = NULL, ..., fill = FALSE,
  reg = getDefaultRegistry())
}
\arguments{
\item{ids}{[\code{\link[base]{data.frame}} or \code{integer}]\cr
A \code{\link[base]{data.frame}} (or \code{\link[data.table]{data.table}})
with a column named \dQuote{job.id}.
This is usually the result of a \code{\link{findJobs}} operation, but you may also
pass any other table here, e.g. the return of \code{\link{getJobInfo}}.
Alternatively, you may also pass a vector of integerish job ids.
If not set, defaults to result of \code{\link{findDone}}.}

\item{fun}{[\code{function}]\cr
Function to apply to each result. The results are passed unnamed as first
argument.}

\item{...}{[\code{ANY}]\cr
Additional arguments passed to to function \code{fun}.}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explictitly passed, uses the last created registry.}

\item{fill}{[\code{logical(1)}]\cr
In \code{reduceResultsDataTable}: This flag is passed down to
\code{\link[data.table]{rbindlist}} which is used to convert the results
to a \code{\link[data.table]{data.table}}.}
}
\value{
\code{reduceResultsList} returns a list, \code{reduceResultsDataTable}
  returns a \code{\link[data.table]{data.table}} with
  columns \dQuote{job.id} and \dQuote{result}.
}
\description{
Functions to conveniently fetch the results from a \code{\link{Registry}}.
}
\seealso{
\code{\link{reduceResults}}.

Other Results: \code{\link{loadResult}},
  \code{\link{reduceResults}}
}

