% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Registry.r
\name{getDefaultRegistry}
\alias{Registry}
\alias{clearDefaultRegistry}
\alias{getDefaultRegistry}
\alias{loadRegistry}
\alias{makeRegistry}
\alias{makeTempRegistry}
\alias{saveRegistry}
\alias{syncRegistry}
\title{Construct a Registry Object}
\usage{
getDefaultRegistry()

clearDefaultRegistry()

makeRegistry(file.dir = "registry", work.dir = getwd(),
  conf.file = getOption("batchtools.conf.file"), packages = character(0L),
  namespaces = character(0L), seed = NULL, make.default = TRUE)

makeTempRegistry(make.default = FALSE,
  temp.dir = getOption("batchtools.temp.dir", tempdir()), ...)

loadRegistry(file.dir = "registry",
  conf.file = getOption("batchtools.conf.file"), make.default = TRUE)

saveRegistry(reg = getDefaultRegistry())

syncRegistry(reg = getDefaultRegistry(), save = TRUE)
}
\arguments{
\item{file.dir}{[\code{character(1)}]\cr
Path where all files of the registry are saved.
Default is directory \dQuote{registry} in the current working directory.}

\item{work.dir}{[\code{character(1)}]\cr
Working directory for R process when experiment is executed.
Default is the current working directory when registry is created.}

\item{conf.file}{[\code{character(1)}]\cr
Path to a configuration file which is sourced directly after the registry is created.
For example, you can set system-specific cluster functions in it.
The script is executed inside the registry environment, thus you can directly set
all slots, e.g. \dQuote{debug = TRUE} would overwrite the debug flag of the registry.
The default location of the configuration can be set via the option \dQuote{batchtools.conf.file}.}

\item{packages}{[\code{character}]\cr
Packages that will always be loaded on each node.
Uses \code{\link[base]{require}} internally.
Default is \code{character(0)}.}

\item{namespaces}{[\code{character}]\cr
Same as \code{packages}, but the packages will not be attached.
Uses \code{\link[base]{requireNamespace}} internally.
Default is \code{character(0)}.}

\item{seed}{[\code{integer(1)}]\cr
Start seed for jobs. Each job uses the (\code{seed} + \code{job.id}) as seed.
Default is a random number in the range [1, \code{.Machine$integer.max/2}].}

\item{make.default}{[\code{logical(1)}]\cr
If set to \code{TRUE}, the created registry is saved inside the package
namespace and acts as default registry. You might want to switch this
off if you work with multiple registries simultaneously.
Default is \code{TRUE}.}

\item{temp.dir}{[\code{character(1)}]\cr
Path to temporary directory.
Defaults to the value of the option \dQuote{batchtools.temp.dir} or (if unset) \code{\link[base]{tempdir}}.}

\item{...}{[\code{ANY}]\cr
Additional parameters passed to \code{makeRegistry}.}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explictitly passed, uses the last created registry.}

\item{save}{[\code{logical(1)}]\cr
Save the registry to the file system? Default is \code{TRUE}.}
}
\value{
[\code{Registry}]. An environment with the following slots:
  \describe{
    \item{file.dir}{File directory.}
    \item{work.dir}{Working directory.}
    \item{packages}{Packages to load on the slaves.}
    \item{namespaces}{Namespaces to load on the slaves.}
    \item{seed}{Registry seed. Before each job is executed, the seed \code{seed + job.id} is set.}
    \item{debug}{Flag to turn additional debug functionality on.}
    \item{cluster.functions}{Usually set in your \code{conf.file}. Set via a call to \code{\link{makeClusterFunctions}}. See example.}
    \item{default.resources}{Usually set in your \code{conf.file}. Named list of default resources.}
    \item{defs}{Table with job definitions. Also see \code{\link{getJobDefs}}.}
    \item{status}{Table holding information about the computational status. Also see \code{\link{getJobStatus}}.}
    \item{resources}{Table holding information about the computational resources used for the job. Also see \code{\link{getJobResources}}.}
  }
}
\description{
\code{makeRegistry} constructs the inter-communication object for all functions in \code{batchtools}.
The registry created last is saved in the package namespace (unless \code{make.default} is set to
\code{FALSE}) and can be retrieved via \code{getDefaultRegistry}.

\code{makeTempRegistry} creates a temporary registry in the subdirectory of a temporary directory.
Such disposable registries are mainly used in sequential apply functions like \code{\link{btlapply}}.
The base dir can be set via the option \dQuote{batchtools.temp.dir} and defaults to \code{\link[base]{tempdir}}.

\code{saveRegistry} serializes the registry to the file system and be be loaded with \code{loadRegistry} by specifying the \code{file.dir}.
\code{syncRegisty} refreshes the registry by parsing update files from jobs.
}
\examples{
reg = makeTempRegistry(make.default = FALSE)
print(reg)

#' Set debug mode
reg$debug = TRUE

# Set cluster functions to interactive mode (default)
reg$cluster.functions = makeClusterFunctionsInteractive()

# Change default packages
reg$packages = c("MASS")
saveRegistry(reg = reg)
}

