% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeExperiments.R
\name{removeExperiments}
\alias{removeExperiments}
\title{Remove Experiments}
\usage{
removeExperiments(ids = NULL, reg = getDefaultRegistry())
}
\arguments{
\item{ids}{[\code{\link[base]{data.frame}} or \code{integer}]\cr
A \code{\link[base]{data.frame}} (or \code{\link[data.table]{data.table}})
with a column named \dQuote{job.id}.
Alternatively, you may also pass a vector of integerish job ids.
If not set, defaults to no job.
Invalid ids are ignored.}

\item{reg}{[\code{\link{ExperimentRegistry}}]\cr
Registry. If not explicitly passed, uses the last created registry.}
}
\value{
[\code{\link{data.table}}] of removed job ids, invisibly.
}
\description{
Remove Experiments from an \code{\link{ExperimentRegistry}}.
This function automatically checks if any of the jobs to reset is either pending or running.
However, if the implemented heuristic fails, this can lead to inconsistencies in the data base.
Use with care while jobs are running.
}
\seealso{
Other Experiment: \code{\link{addExperiments}},
  \code{\link{summarizeExperiments}}
}
\concept{Experiment}
