% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterFunctions.R
\name{runOSCommand}
\alias{runOSCommand}
\title{Run OS Commands on local or remote machines}
\usage{
runOSCommand(sys.cmd, sys.args = character(0L), nodename = "localhost",
  stop.on.exit.code = TRUE, debug = FALSE)
}
\arguments{
\item{sys.cmd}{[\code{character(1)}]\cr
Command to run.}

\item{sys.args}{[\code{character()}]\cr
Arguments to \code{sys.cmd}.}

\item{nodename}{[\code{character(1)}]\cr
Name of the SSH node to run the command on. If set to \dQuote{localhost} (default), the command
is not piped through SSH.}

\item{stop.on.exit.code}{[\code{logical(1)}]\cr
Throw an error message if the exit code of \code{sys.cmd} is not \dQuote{0}?}

\item{debug}{[\code{logical(1)}]\cr
If set to \code{TRUE}, prints the complete command, exit code and output.}
}
\value{
[\code{named list}] with \dQuote{exit.code} (integer) and \dQuote{output} (character).
}
\description{
This is a helper function to run arbitrary OS commands on local or remote machines.
The interface is similar to \code{\link[base]{system2}}, but it always returns the exit status
and the output.
}
\examples{
\dontrun{
runOSCommand("ls")
runOSCommand("ls", "-al")
runOSCommand("notfound", stop.on.exit.code = FALSE)
}
}
\seealso{
Other ClusterFunctions: \code{\link{makeClusterFunctionsDocker}},
  \code{\link{makeClusterFunctionsLSF}},
  \code{\link{makeClusterFunctionsOpenLava}},
  \code{\link{makeClusterFunctionsSGE}},
  \code{\link{makeClusterFunctionsSLURM}},
  \code{\link{makeClusterFunctionsTorque}},
  \code{\link{makeClusterFunctions}}
}

