% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveRegistry.R
\name{saveRegistry}
\alias{saveRegistry}
\title{Store the Registy to the File System}
\usage{
saveRegistry(reg = getDefaultRegistry())
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explicitly passed, uses the default registry (see \code{\link{setDefaultRegistry}}).}
}
\value{
[\code{logical(1)}]: \code{TRUE} if the registry was saved,
  \code{FALSE} otherwise (if the registry is read-only).
}
\description{
Stores the registry on the file system in its \dQuote{file.dir} (specified
for construction in \code{\link{makeRegistry}}, can be accessed via
\code{reg$file.dir}).
This function is usually called internally whenever needed.
}
\seealso{
Other Registry: \code{\link{clearRegistry}},
  \code{\link{getDefaultRegistry}},
  \code{\link{loadRegistry}}, \code{\link{makeRegistry}},
  \code{\link{removeRegistry}},
  \code{\link{sweepRegistry}}, \code{\link{syncRegistry}}
}
\concept{Registry}
