% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobNames.R
\name{setJobNames}
\alias{setJobNames}
\title{Set Custom Job Names}
\usage{
setJobNames(ids, names, reg = getDefaultRegistry())
}
\arguments{
\item{ids}{[\code{\link[base]{data.frame}} or \code{integer}]\cr
A \code{\link[base]{data.frame}} (or \code{\link[data.table]{data.table}})
with a column named \dQuote{job.id}.
Alternatively, you may also pass a vector of integerish job ids.
If not set, defaults to no job.
Invalid ids are ignored.}

\item{names}{[\code{character}]\cr
Character vector of the same length as provided ids.}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explicitly passed, uses the default registry (see \code{\link{setDefaultRegistry}}).}
}
\value{
Nothing.
}
\examples{
tmp = makeRegistry(file.dir = NA, make.default = FALSE)
ids = batchMap(identity, 1:10, reg = tmp)
setJobNames(ids, letters[1:nrow(ids)], reg = tmp)
}
