% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logs.R
\name{showLog}
\alias{getLog}
\alias{showLog}
\title{Inspect Log Files}
\usage{
showLog(id, reg = getDefaultRegistry())

getLog(id, reg = getDefaultRegistry())
}
\arguments{
\item{id}{[\code{integer(1)} or \code{data.table}]\cr
Single integer to specify the job or a \code{data.table} with column \code{job.id}
and exactly one row.}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explicitly passed, uses the last created registry.}
}
\value{
Nothing.
}
\description{
\code{showLog} opens the log in the pager. For customization, see \code{\link[base]{file.show}}.
\code{getLog} returns the log as character vector.
}
\seealso{
Other debug: \code{\link{getErrorMessages}},
  \code{\link{grepLogs}}, \code{\link{killJobs}},
  \code{\link{testJob}}
}

