% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarizeExperiments.r
\name{summarizeExperiments}
\alias{summarizeExperiments}
\title{Quick Summary over Experiments}
\usage{
summarizeExperiments(ids = NULL, by = c("problem", "algorithm"),
  reg = getDefaultRegistry())
}
\arguments{
\item{ids}{[\code{\link[base]{data.frame}} or \code{integer}]\cr
A \code{\link[base]{data.frame}} (or \code{\link[data.table]{data.table}})
with a column named \dQuote{job.id}.
This is usually the result of a \code{\link{findJobs}} operation, but you may also
pass any other table here, e.g. the return of \code{\link{getJobInfo}}.
Alternatively, you may also pass a vector of integerish job ids.
If not set, defaults to all jobs.}

\item{by}{[\code{character}]\cr
Split the resulting table by these columns of \code{\link{getJobDefs}}.}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explictitly passed, uses the last created registry.}
}
\value{
[\code{\link[data.table]{data.table}}] of frequencies in the respective groups.
}
\description{
Returns a frequency table of defined experiments.
}

