% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepRegistry.R
\name{sweepRegistry}
\alias{sweepRegistry}
\title{Check Consistency and Remove Obsolete Information}
\usage{
sweepRegistry(reg = getDefaultRegistry())
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explicitly passed, uses the default registry (see \code{\link{setDefaultRegistry}}).}
}
\description{
Canceled jobs and jobs submitted multiple times may leave stray files behind.
This function checks the registry for consistency and removes obsolete files
and data base informations.
}
\seealso{
Other Registry: \code{\link{clearRegistry}},
  \code{\link{getDefaultRegistry}},
  \code{\link{loadRegistry}}, \code{\link{makeRegistry}},
  \code{\link{removeRegistry}}, \code{\link{saveRegistry}},
  \code{\link{syncRegistry}}
}
