% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncRegistry.R
\name{syncRegistry}
\alias{syncRegistry}
\title{Syncronize the Registry}
\usage{
syncRegistry(reg = getDefaultRegistry())
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explicitly passed, uses the default registry (see \code{\link{setDefaultRegistry}}).}
}
\value{
[\code{logical(1)}]: \code{TRUE} if the state has changed, \code{FALSE} otherwise.
}
\description{
Parses update files written by the slaves to the file system and updates the
internal data base.
}
\seealso{
Other Registry: \code{\link{clearRegistry}},
  \code{\link{getDefaultRegistry}},
  \code{\link{loadRegistry}}, \code{\link{makeRegistry}},
  \code{\link{removeRegistry}}, \code{\link{saveRegistry}},
  \code{\link{sweepRegistry}}
}
