% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tags.R
\name{Tags}
\alias{Tags}
\alias{addJobTags}
\alias{getUsedJobTags}
\alias{removeJobTags}
\title{Add or Remove Job Tags}
\usage{
addJobTags(ids = NULL, tags, reg = getDefaultRegistry())

removeJobTags(ids = NULL, tags, reg = getDefaultRegistry())

getUsedJobTags(ids = NULL, reg = getDefaultRegistry())
}
\arguments{
\item{ids}{[\code{\link[base]{data.frame}} or \code{integer}]\cr
A \code{\link[base]{data.frame}} (or \code{\link[data.table]{data.table}})
with a column named \dQuote{job.id}.
Alternatively, you may also pass a vector of integerish job ids.
If not set, defaults to all jobs.}

\item{tags}{[\code{character}]\cr
Tags to add or remove as string. May use letters, numbers, underscore and dots (pattern \dQuote{^[[:alnum:]_.]+}).}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explicitly passed, uses the last created registry.}
}
\value{
[\code{\link[data.table]{data.table}}] with job ids affected (invisible).
}
\description{

}
\examples{
reg = makeRegistry(file.dir = NA, make.default = FALSE)
ids = batchMap(sqrt, x = -3:3, reg = reg)

# Add new tag to all ids
addJobTags(ids, "needs.computation", reg = reg)
getJobTags(reg = reg)

# Add more tags
addJobTags(findJobs(x < 0, reg = reg), "x.neg", reg = reg)
addJobTags(findJobs(x > 0, reg = reg), "x.pos", reg = reg)
getJobTags(reg = reg)

# Submit first 5 jobs and remove tag if successful
ids = submitJobs(1:5, reg = reg)
if (waitForJobs(reg = reg))
  removeJobTags(ids, "needs.computation", reg = reg)
getJobTags(reg = reg)

# Grep for warning message and add a tag
addJobTags(grepLogs(pattern = "NaNs produced", reg = reg), "div.zero", reg = reg)
getJobTags(reg = reg)

# All tags where tag x.neg is set:
ids = findTagged("x.neg", reg = reg)
getUsedJobTags(ids, reg = reg)
}

