% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testJob.r
\name{testJob}
\alias{testJob}
\title{Run Jobs Interactively}
\usage{
testJob(id, reg = getDefaultRegistry())
}
\arguments{
\item{id}{[\code{integer(1)} or \code{data.table}]\cr
Single integer to specify the job or a \code{data.table} with column \code{job.id}
and exactly one row.}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explicitly passed, uses the last created registry.}
}
\value{
[ANY]. Returns the result of the job if successful.
}
\description{
Starts a single job in the current R session.
Debugging tools like \code{\link[base]{traceback}} should work.
}
\examples{
reg = makeTempRegistry(make.default = FALSE)
batchMap(function(x) if (x == 2) stop(x) else x, 1:2, reg = reg)
testJob(1, reg = reg)
\dontrun{
 testJob(2, reg = reg)
}
}
\seealso{
Other debug: \code{\link{killJobs}}, \code{\link{showLog}}
}

