% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waitForJobs.r
\name{waitForJobs}
\alias{waitForJobs}
\title{Wait for termination of jobs on the batch system}
\usage{
waitForJobs(ids = NULL, sleep = 10, timeout = 604800,
  stop.on.error = FALSE, reg = getDefaultRegistry())
}
\arguments{
\item{ids}{[\code{\link[base]{data.frame}} or \code{integer}]\cr
A \code{\link[base]{data.frame}} (or \code{\link[data.table]{data.table}})
with a column named \dQuote{job.id}.
This is usually the result of a \code{\link{findJobs}} operation, but you may also
pass any other table here, e.g. the return of \code{\link{getJobInfo}}.
Alternatively, you may also pass a vector of integerish job ids.
If not set, defaults to result of \code{\link{findNotDone}}.}

\item{sleep}{[\code{numeric(1)}]\cr
Seconds to sleep between status updates. Default is \code{10}.}

\item{timeout}{[\code{numeric(1)}]\cr
After waiting \code{timeout} seconds, show a message and return
\code{FALSE}. This argument may be required on some systems where, e.g.,
expired jobs or jobs on hold are problematic to detect. If you don't want
a timeout, set this to \code{Inf}. Default is \code{604800} (one week).}

\item{stop.on.error}{[\code{logical(1)}]\cr
Immediately cancel if a job terminates with an error? Default is
\code{FALSE}.}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explictitly passed, uses the last created registry.}
}
\value{
[\code{logical(1)}]. Returns \code{TRUE} if all jobs terminated
successfully and \code{FALSE} if either the timeout is reached or at least
one job terminated with an exception.
}
\description{
This function can be used to synchronise the execution on batch systems.
}

