% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Worker.r
\name{Worker}
\alias{Worker}
\alias{makeWorker}
\title{Worker Constructor}
\usage{
makeWorker(nodename, ncpus = 0L, max.load = Inf, debug = FALSE)
}
\arguments{
\item{nodename}{[\code{character(1)}]\cr
Host name of node.}

\item{ncpus}{[\code{integers(1)}]\cr
Number of VPUs of worker. Default (0) means to query the worker.}

\item{max.load}{[\code{numeric(1)}]\cr
Load average (of the last 5 min) at which the worker is considered
occupied, so that no job can be submitted.
Default is \code{Inf}.}

\item{debug}{[\code{logical(1)}]\cr
Print some verbose info messages. Default is \code{FALSE}.}
}
\value{
[\code{\link{Worker}}].
}
\description{
Functions to construct local and remote workers.
}
\examples{
\dontrun{
# create a worker for the local machine and use 4 CPUs.
makeWorker("localhost", ncpus = 4)
}
}

