% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{batchtools-package}
\alias{batchtools}
\alias{batchtools-package}
\title{Tools for Computation on Batch Systems}
\description{
For bug reports and feature requests please use the tracker:
\url{https://github.com/mllg/batchtools}.
}
\section{Package options}{

\describe{
  \item{\code{batchtools.progress}}{
    Progress bars. Set to \code{FALSE} to disable them.
  }
  \item{\code{batchtools.verbose}}{
    Verbosity. Set to \code{FALSE} to suppress info messages and progress bars.
  }
}
Furthermore, you may enable a debug mode using the \pkg{debugme} package by
setting the environment variable \dQuote{DEBUGME} to \dQuote{batchtools} before
loading \pkg{batchtools}.
}

