% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{batchtools-package}
\alias{batchtools}
\alias{batchtools-package}
\title{batchtools: Tools for Computation on Batch Systems}
\description{
For bug reports and feature requests please use the tracker:
\url{https://github.com/mllg/batchtools}.
}
\section{Package options}{

\describe{
  \item{\code{batchtools.progress}}{
    Progress bars. Set to \code{FALSE} to disable them.
  }
  \item{\code{batchtools.verbose}}{
    Verbosity. Set to \code{FALSE} to suppress info messages and progress bars.
  }
}
Furthermore, you may enable a debug mode using the \pkg{debugme} package by
setting the environment variable \dQuote{DEBUGME} to \dQuote{batchtools} before
loading \pkg{batchtools}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mllg/batchtools}
  \item Report bugs at \url{https://github.com/mllg/batchtools/issues}
}

}
\author{
\strong{Maintainer}: Michel Lang \email{michellang@gmail.com}

Authors:
\itemize{
  \item Bernd Bischl \email{bernd_bischl@gmx.de}
}

Other contributors:
\itemize{
  \item Dirk Surmann \email{surmann@statistik.tu-dortmund.de} [contributor]
}

}
