% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btlapply.R
\name{btlapply}
\alias{btlapply}
\alias{btmapply}
\title{Synchronous Apply Functions}
\usage{
btlapply(X, fun, ..., reg = makeTempRegistry(make.default = FALSE))

btmapply(fun, ..., more.args = list(), simplify = FALSE, use.names = TRUE,
  reg = makeTempRegistry(make.default = FALSE))
}
\arguments{
\item{X}{[\code{\link[base]{vector}}]\cr
Vector to apply over.}

\item{fun}{[\code{function}]\cr
Function to apply.}

\item{...}{[\code{ANY}]\cr
Additional arguments passed to \code{fun} (\code{btlapply}) or vectors to map over (\code{btmapply}).}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explicitly passed, uses the last created registry.}

\item{more.args}{[\code{list}]\cr
Additional arguments passed to \code{fun}.}

\item{simplify}{[\code{logical(1)}]\cr
Simplify the results using \code{\link[base]{simplify2array}}?}

\item{use.names}{[\code{logical(1)}]\cr
Use names of the input to name the output?}
}
\value{
[\code{list}] List with the results of the function call.
}
\description{
This is a set of convenience functions as counterparts to the sequential
apply functions in base R: \code{btlapply} for \code{\link[base]{lapply}} and
\code{btmapply} for \code{link[base]{mapply}}.
Internally the jobs are created in a temporary registry (see \code{\link{makeTempRegistry}}
and \code{\link{batchMap}} is called on the input vector(s).
After \code{\link{waitForJobs}} terminated, the result is reduced into a list.
Because the result is only returned as soon as all jobs are terminated,
the execution is called synchronized (in contrast to the usual asynchronous execution on
batch systems).

Note that these functions are one suitable for short and fail-safe operations
on batch system. If some jobs fail, you have to retrieve partial results from the
temporary registry directory yourself.
}

