% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clusterFunctions.r
\name{cfBrewTemplate}
\alias{cfBrewTemplate}
\title{Cluster functions helper: Brew your template into a job description file}
\usage{
cfBrewTemplate(reg, template, jc)
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explictitly passed, uses the last created registry.}

\item{template}{[\code{character(1)}]\cr
String with contents of the brew templated, as returned by \code{\link{cfReadBrewTemplate}}.}

\item{jc}{[\code{\link{JobCollection})}]\cr
JobCollection holding all essential information. Will be used as environment to look
up variables.}
}
\value{
[\code{character(1)}]. File path to resulting template file.
}
\description{
This function is only intended for use in your own cluster functions implementation.

Calls brew silently on your template, any error will lead to an exception. If debug mode is
turned on, the file is stored at the same place as the corresponding R script in the
\dQuote{jobs}-subdir of your files directory, otherwise in the temp dir via
\code{\link{tempfile}}.
}
\seealso{
Other ClusterFunctionsHelper: \code{\link{SubmitJobResult}},
  \code{\link{makeSubmitJobResult}};
  \code{\link{cfHandleUnknownSubmitError}};
  \code{\link{cfKillBatchJob}};
  \code{\link{cfReadBrewTemplate}}
}

