% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterFunctions.r
\name{cfReadBrewTemplate}
\alias{cfReadBrewTemplate}
\title{Cluster functions helper: Read in your brew template file}
\usage{
cfReadBrewTemplate(template, comment.string = NA_character_)
}
\arguments{
\item{template}{[\code{character(1)}]\cr
Path to a brew template file that is used for the job description.}

\item{comment.string}{[\code{character(1)}]\cr
Ignore lines starting with this string.}
}
\value{
[\code{character}].
}
\description{
This function is only intended for use in your own cluster functions implementation.

Simply reads your template and returns it as a character vector. If you do this in the
constructor of your cluster functions once, you can avoid this repeated file access later on.
}
\seealso{
Other ClusterFunctionsHelper: \code{\link{cfBrewTemplate}},
  \code{\link{cfHandleUnknownSubmitError}},
  \code{\link{cfKillBatchJob}},
  \code{\link{makeSubmitJobResult}}
}

