% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterFunctions.R
\name{cfReadBrewTemplate}
\alias{cfReadBrewTemplate}
\title{Cluster Functions Helper to Parse a Brew Template}
\usage{
cfReadBrewTemplate(template, comment.string = NA_character_)
}
\arguments{
\item{template}{[\code{character(1)}]\cr
Path to template file or single string (containing newlines) which is then passed
to \code{\link[brew]{brew}}.}

\item{comment.string}{[\code{character(1)}]\cr
Ignore lines starting with this string.}
}
\value{
[\code{character}].
}
\description{
This function is only intended for use in your own cluster functions implementation.

This function is only intended for use in your own cluster functions implementation.
Simply reads your template file and returns it as a character vector.
}
\seealso{
Other ClusterFunctionsHelper: \code{\link{cfBrewTemplate}},
  \code{\link{cfHandleUnknownSubmitError}},
  \code{\link{cfKillJob}},
  \code{\link{makeClusterFunctions}},
  \code{\link{makeSubmitJobResult}},
  \code{\link{runOSCommand}}
}
