% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doJobCollection.r
\name{doJobCollection}
\alias{doJobCollection}
\title{Execute Jobs of a JobCollection}
\usage{
doJobCollection(jc)
}
\arguments{
\item{jc}{[\code{\link{JobCollection}}]\cr
Either an object of class \dQuote{JobCollection} as returned by
\code{\link{makeJobCollection}} or a string point to file containing a
\dQuote{JobCollection} (saved with \code{\link[base]{saveRDS}}).}
}
\value{
[\code{data.table}]. Data table with updates on the computational
status, i.e. a table with the columns \dQuote{job.id}, \dQuote{started}
(unix time stamp of job start), \dQuote{done} (unix time stamp of job
termination), \dQuote{error} (error message as string) and \code{memory}
(memory usage as double).
}
\description{
Executes every job in a \code{\link{JobCollection}}.
This function is intended to be called on the slave.
}

