% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execJob.R
\name{execJob}
\alias{execJob}
\title{Execute a Single Jobs}
\usage{
execJob(job)
}
\arguments{
\item{job}{[\code{\link{Job}} | \code{\link{Experiment}}]\cr
Job/Experiment to execute.}
}
\value{
Result of the job.
}
\description{
Executes a single job (as created by \code{\link{makeJob}}) and returns
its result. Also works for Experiments.
}
\examples{
\dontshow{ batchtools:::example_push_temp(1) }
tmp = makeRegistry(file.dir = NA, make.default = FALSE)
batchMap(identity, 1:2, reg = tmp)
job = makeJob(1, reg = tmp)
execJob(job)
}
