% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findJobs.r
\name{findJobs}
\alias{findDone}
\alias{findError}
\alias{findExpired}
\alias{findJobs}
\alias{findNotDone}
\alias{findNotStarted}
\alias{findNotSubmitted}
\alias{findOnSystem}
\alias{findStarted}
\alias{findSubmitted}
\title{Find and filter jobs}
\usage{
findJobs(expr, ids = NULL, reg = getDefaultRegistry())

findSubmitted(ids = NULL, reg = getDefaultRegistry())

findNotSubmitted(ids = NULL, reg = getDefaultRegistry())

findStarted(ids = NULL, reg = getDefaultRegistry())

findNotStarted(ids = NULL, reg = getDefaultRegistry())

findDone(ids = NULL, reg = getDefaultRegistry())

findNotDone(ids = NULL, reg = getDefaultRegistry())

findError(ids = NULL, reg = getDefaultRegistry())

findOnSystem(ids = NULL, reg = getDefaultRegistry())

findExpired(ids = NULL, reg = getDefaultRegistry())
}
\arguments{
\item{expr}{[\code{expression}]\cr
All jobs for which the expression evaluates to \code{TRUE} are selected.}

\item{ids}{[\code{\link[base]{data.frame}} or \code{integer}]\cr
A \code{\link[base]{data.frame}} (or \code{\link[data.table]{data.table}})
with a column named \dQuote{job.id}.
This is usually the result of a \code{\link{findJobs}} operation, but you may also
pass any other table here, e.g. the return of \code{\link{getJobInfo}}.
Alternatively, you may also pass a vector of integerish job ids.
If not set, defaults to all jobs.}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explictitly passed, uses the last created registry.}
}
\value{
[\code{data.table}]. Matching job ids are stored in the column \dQuote{job.id}.
}
\description{
Use \code{findJobs} to query jobs for which a predicate expression, evaluated on the parameters, yields \code{TRUE}.
The other functions can be used to query the computational status.
Note that they do not synchronize the registry, thus you are advised to run \code{\link{syncRegistry}} yourself
or, if you are really just interested in the status, use \code{\link{getStatus}}.
}
\examples{
reg = makeTempRegistry(make.default = FALSE)
batchMap(identity, i = 1:3, reg = reg)
ids = findNotSubmitted(reg = reg)

# get all jobs:
findJobs(reg = reg)

# filter for jobs with parameter i >= 2
findJobs(i >= 2, reg = reg)

# filter on the computational status
findSubmitted(reg = reg)
findNotDone(reg = reg)
}

