% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{JoinTables}
\alias{JoinTables}
\alias{ajoin}
\alias{ijoin}
\alias{ljoin}
\alias{ojoin}
\alias{rjoin}
\alias{sjoin}
\title{Inner, Left, Right, Outer and Anti Join for Job Tables}
\usage{
ijoin(x, y)

ljoin(x, y)

rjoin(x, y)

ojoin(x, y)

sjoin(x, y)

ajoin(x, y)
}
\arguments{
\item{x}{[\code{\link{data.table}}]\cr
Data table with key \dQuote{job.id} as returned by most functions in batchtools.}

\item{y}{[\code{\link{data.table}}]\cr
Data table with key \dQuote{job.id} as returned by most functions in batchtools.}
}
\value{
[\code{\link{data.table}}] with key \dQuote{job.id}.
  See \code{\link{JoinTables}} for examples on working with job tables.
}
\description{
These helper functions perform join operations on job tables.
They are basically one-liners with additional argument checks for sanity.
See \url{http://rpubs.com/ronasta/join_data_tables} for a overview of join operations in
data table or alternatively \pkg{dplyr}'s vignette on two table verbs.
}
\examples{
# create two tables for demonstration
reg = makeRegistry(file.dir = NA, make.default = FALSE)
batchMap(identity, x = 1:6, reg = reg)
x = getJobPars(reg = reg)
y = findJobs(x >= 2 & x <= 5, reg = reg)
y$extra.col = head(letters, nrow(y))

# inner join: similar to intersect() on ids, keep all columns of x and y
ijoin(x, y)

# left join: use all ids from x, keep all columns of x and y
ljoin(x, y)

# right join: use all ids from y, keep all columns of x and y
rjoin(x, y)

# outer join: similar to union() on ids, keep all columns of x and y
ojoin(x, y)

# semi join: similar to intersect() on ids, keep all columns of x
sjoin(x, y)

# anti join: similar to setdiff() on ids, keep all columns of x
ajoin(x, y)
}

