% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Joins.R
\name{JoinTables}
\alias{JoinTables}
\alias{ajoin}
\alias{ijoin}
\alias{ljoin}
\alias{ojoin}
\alias{rjoin}
\alias{sjoin}
\title{Inner, Left, Right, Outer, Semi and Anti Join for Job Tables}
\usage{
ijoin(x, y)

ljoin(x, y)

rjoin(x, y)

ojoin(x, y)

sjoin(x, y)

ajoin(x, y)
}
\arguments{
\item{x}{[\code{\link{data.frame}} | \code{integer}]\cr
Either a \code{\link[data.table]{data.table}}/\code{\link[base]{data.frame}} with integer column \dQuote{job.id}
or an integer vector of job ids.}

\item{y}{[\code{\link{data.frame}} | \code{integer}]\cr
Either a \code{\link[data.table]{data.table}}/\code{\link[base]{data.frame}} with integer column \dQuote{job.id}
or an integer vector of job ids.}
}
\value{
[\code{\link{data.table}}] with key \dQuote{job.id}.
}
\description{
These helper functions perform join operations on job tables.
They are basically one-liners with additional argument checks for sanity.
See \url{http://rpubs.com/ronasta/join_data_tables} for a overview of join operations in
data table or alternatively \pkg{dplyr}'s vignette on two table verbs.
}
\examples{
# Create two tables for demonstration
tmp = makeRegistry(file.dir = NA, make.default = FALSE)
batchMap(identity, x = 1:6, reg = tmp)
x = getJobPars(reg = tmp)
y = findJobs(x >= 2 & x <= 5, reg = tmp)
y$extra.col = head(letters, nrow(y))

# Inner join: similar to intersect() on ids, keep all columns of x and y
ijoin(x, y)

# Left join: use all ids from x, keep all columns of x and y
ljoin(x, y)

# Right join: use all ids from y, keep all columns of x and y
rjoin(x, y)

# Outer join: similar to union() on ids, keep all columns of x and y
ojoin(x, y)

# Semi join: similar to intersect() on ids, keep all columns of x
sjoin(x, y)

# Anti join: similar to setdiff() on ids, keep all columns of x
ajoin(x, y)
}

