% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResult.R
\name{loadResult}
\alias{loadResult}
\title{Load the Result of a Single Job}
\usage{
loadResult(id, reg = getDefaultRegistry())
}
\arguments{
\item{id}{[\code{integer(1)} or \code{data.table}]\cr
Single integer to specify the job or a \code{data.table} with column \code{job.id}
and exactly one row.}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explicitly passed, uses the default registry (see \code{\link{setDefaultRegistry}}).}
}
\value{
[\code{ANY}]. The stored result.
}
\description{
Loads the result of a single job.
}
\seealso{
Other Results: 
\code{\link{batchMapResults}()},
\code{\link{reduceResultsList}()},
\code{\link{reduceResults}()}
}
\concept{Results}
