% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterFunctionsLSF.R
\name{makeClusterFunctionsLSF}
\alias{makeClusterFunctionsLSF}
\title{ClusterFunctions for LSF Systems}
\usage{
makeClusterFunctionsLSF(template = findTemplateFile("lsf"),
  scheduler.latency = 1, fs.latency = 65)
}
\arguments{
\item{template}{[\code{character(1)}]\cr
Path to a brew template file that can be used for the job description.
Alternatively, the template as a single string (including at least one newline \dQuote{\\n}).
Defaults to a heuristic which looks for a template file in the following locations and picks
the first one found:
\enumerate{
  \item \dQuote{batchtools.lsf.tmpl} in the current working directory.
  \item \dQuote{lsf.tmpl} in the user config directory (see \code{\link[rappdirs]{user_config_dir}}); on linux is is usually \dQuote{~/.config/batchtools/lsf.tmpl}.
  \item \dQuote{.batchtools.lsf.tmpl} in the home directory.
  \item \dQuote{lsf.default.tmpl} in the package installation directory in the subfolder \dQuote{templates} (this probably does not exist for your cluster system or needs adaptation).
}}

\item{scheduler.latency}{[\code{numeric(1)}]\cr
Time to sleep after important interactions with the scheduler to ensure a sane state.
Currently only triggered after calling \code{\link{submitJobs}}.}

\item{fs.latency}{[\code{numeric(1)}]\cr
Expected maximum latency of the file system, in seconds.
Set to a positive number for network file systems like NFS which enables more robust (but also more expensive) mechanisms to
access files and directories.
Usually safe to set to \code{NA} which disables the expensive heuristic if you are working on a local file system.}
}
\value{
[\code{\link{ClusterFunctions}}].
}
\description{
Cluster functions for LSF (\url{http://www-03.ibm.com/systems/spectrum-computing/products/lsf/}).

Job files are created based on the brew template \code{template.file}. This
file is processed with brew and then submitted to the queue using the
\code{bsub} command. Jobs are killed using the \code{bkill} command and the
list of running jobs is retrieved using \code{bjobs -u $USER -w}. The user
must have the appropriate privileges to submit, delete and list jobs on the
cluster (this is usually the case).

The template file can access all resources passed to \code{\link{submitJobs}}
as well as all variables stored in the \code{\link{JobCollection}}.
It is the template file's job to choose a queue for the job and handle the desired resource
allocations.
}
\note{
Array jobs are currently not supported.
}
\seealso{
Other ClusterFunctions: \code{\link{makeClusterFunctionsDocker}},
  \code{\link{makeClusterFunctionsInteractive}},
  \code{\link{makeClusterFunctionsMulticore}},
  \code{\link{makeClusterFunctionsOpenLava}},
  \code{\link{makeClusterFunctionsSGE}},
  \code{\link{makeClusterFunctionsSSH}},
  \code{\link{makeClusterFunctionsSlurm}},
  \code{\link{makeClusterFunctionsSocket}},
  \code{\link{makeClusterFunctionsTORQUE}},
  \code{\link{makeClusterFunctions}}
}
