% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterFunctionsMulticore.R
\name{makeClusterFunctionsMulticore}
\alias{makeClusterFunctionsMulticore}
\title{ClusterFunctions for Parallel Execution}
\usage{
makeClusterFunctionsMulticore(ncpus = max(getOption("mc.cores",
  parallel::detectCores()), 1L))
}
\arguments{
\item{ncpus}{[\code{integer(1)}]\cr
Number of VPUs of worker.
Default is to use all cores but one, where total number of cores "available" is given by option \code{mc.cores}
and defaults to the heuristic implemented in \code{\link[parallel]{detectCores}}.}
}
\value{
[\code{\link{ClusterFunctions}}].
}
\description{
Jobs are spawned asynchronously using the packages \pkg{parallel}.
Does not work on Windows, use \code{\link{makeClusterFunctionsSocket}} instead.
}
\note{
Sets the number of threads internally used by \pkg{data.table} to 1 during initialization (via \code{data.table::setthreads}).
}
\seealso{
Other clusterFunctions: \code{\link{makeClusterFunctionsInteractive}},
  \code{\link{makeClusterFunctionsSSH}},
  \code{\link{makeClusterFunctionsSocket}}
}

