% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterFunctionsMulticore.r
\name{makeClusterFunctionsMulticore}
\alias{makeClusterFunctionsMulticore}
\title{Cluster Functions for Local Multicore Execution}
\usage{
makeClusterFunctionsMulticore(ncpus = max(getOption("mc.cores",
  parallel::detectCores()) - 1L, 1L), max.jobs, max.load)
}
\arguments{
\item{ncpus}{[\code{integer(1)}]\cr
Number of VPUs of worker.
Default is to use all cores but one, where total number of cores
"available" is given by option \code{\link[base:options]{mc.cores}}
and if that is not set it is inferred by
\code{\link[parallel]{detectCores}}.}

\item{max.jobs}{[\code{integer(1)}]\cr
Maximal number of jobs that can run concurrently for the current registry.
Default is \code{ncpus}.}

\item{max.load}{[\code{numeric(1)}]\cr
Load average (of the last 5 min) at which the worker is considered occupied,
so that no job can be submitted.
Default is inferred by \code{\link[parallel]{detectCores}}, cf. argument \code{ncpus}.}
}
\value{
[\code{\link{ClusterFunctions}}].
}
\description{
Jobs are spawned by starting multiple R sessions on the command line.
}
\seealso{
Other clusterFunctions: \code{\link{makeClusterFunctionsInteractive}}
}

