% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterFunctionsMulticore.R
\name{makeClusterFunctionsMulticore}
\alias{makeClusterFunctionsMulticore}
\title{ClusterFunctions for Local Multicore Execution}
\usage{
makeClusterFunctionsMulticore(ncpus = max(getOption("mc.cores",
  parallel::detectCores()), 1L), max.load = Inf)
}
\arguments{
\item{ncpus}{[\code{integer(1)}]\cr
Number of VPUs of worker.
Default is to use all cores but one, where total number of cores "available" is given by option \code{mc.cores}
and defaults to the heuristic implemented in \code{\link[parallel]{detectCores}}.}

\item{max.load}{[\code{numeric(1)}]\cr
Load average (of the last 5 min) at which the worker is considered occupied,
so that no job can be submitted. Default is \code{Inf}.}
}
\value{
[\code{\link{ClusterFunctions}}].
}
\description{
Jobs are spawned by starting multiple R sessions on the command line.
}
\seealso{
Other clusterFunctions: \code{\link{makeClusterFunctionsInteractive}},
  \code{\link{makeClusterFunctionsSSH}}
}

