% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clusterFunctionsSLURM.r
\name{makeClusterFunctionsSLURM}
\alias{makeClusterFunctionsSLURM}
\title{Cluster Functions for SLURM}
\usage{
makeClusterFunctionsSLURM(template)
}
\arguments{
\item{template}{[\code{character(1)}]\cr
Path to a brew template file that is used for the job description.}
}
\value{
[\code{\link{ClusterFunctions}}].
}
\description{
Job files are created based on the brew template \code{template.file}. This
file is processed with brew and then submitted to the queue using the
\code{sbatch} command. Jobs are killed using the \code{scancel} command and
the list of running jobs is retrieved using \code{squeue}. The user must
have the appropriate privileges to submit, delete and list jobs on the
cluster (this is usually the case).

The template file can access all arguments passed to the \code{submitJob}
function, see here \code{\link{ClusterFunctions}}. It is the template file's
job to choose a queue for the job and handle the desired resource
allocations.
}
\seealso{
Other ClusterFunctions: \code{\link{ClusterFunctions}},
  \code{\link{makeClusterFunctions}};
  \code{\link{makeClusterFunctionsLSF}};
  \code{\link{makeClusterFunctionsOpenLava}};
  \code{\link{makeClusterFunctionsSGE}};
  \code{\link{makeClusterFunctionsTorque}}
}

