% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Job.r
\name{makeJob}
\alias{Job}
\alias{makeJob}
\title{Computational Jobs}
\usage{
makeJob(id, reg = getDefaultRegistry())
}
\arguments{
\item{id}{[\code{integer(1)} or \code{data.table}]\cr
Single integer to specify the job or a \code{data.table} with column \code{job.id}
and exactly one row.}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explicitly passed, uses the last created registry.}
}
\value{
[\code{Job}].
}
\description{
\code{makeJob} takes a single job id and creates an object of class \dQuote{Job}. This job is passed to reduce
functions like \code{\link{reduceResults}}. Furthermore, it can be used for an \code{\link{ExperimentRegistry}} while
in the user functions of each \code{\link{Problem}} and \code{\link{Algorithm}}. It is implemented as a list holding
the following variables:
\describe{
 \item{job.id}{Job ID as integer.}
 \item{pars}{Job parameters as named list. For \code{\link{ExperimentRegistry}}, the parameters are divided into the
   sublists \dQuote{prob.pars} and \dQuote{algo.pars}.}
 \item{seed}{Seed which is set via \code{\link{doJobCollection}}.}
 \item{resources}{Computational resources which were set for this job.}
}
Additionally, a regular \code{\link{Registry}} also has the slot \dQuote{fun} with the user function while an
\code{\link{ExperimentRegistry}} has the slots \dQuote{problem}, \dQuote{algorithm} and \dQuote{repl} (replication).

Jobs can be executed with \code{\link{execJob}}.
}
\examples{
reg = makeTempRegistry(make.default = FALSE)
batchMap(identity, 1:5, reg = reg)
job = makeJob(1, reg = reg)
names(job)
}

