% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Results.r
\name{reduceResults}
\alias{reduceResults}
\title{Reduce Results}
\usage{
reduceResults(fun, ids = NULL, init, ..., reg = getDefaultRegistry())
}
\arguments{
\item{fun}{[\code{function}]\cr
A function to reduce the results. The current result will be passed unnamed
as first argument. See \code{\link[base]{Reduce}} for some examples.}

\item{ids}{[\code{\link[base]{data.frame}} or \code{integer}]\cr
A \code{\link[base]{data.frame}} (or \code{\link[data.table]{data.table}})
with a column named \dQuote{job.id}.
This is usually the result of a \code{\link{findJobs}} operation, but you may also
pass any other table here, e.g. the return of \code{\link{getJobInfo}}.
Alternatively, you may also pass a vector of integerish job ids.
If not set, defaults to result of \code{\link{findDone}}.}

\item{init}{[\code{ANY}]\cr
Initial element, as used in \code{\link[base]{Reduce}}.
Default is the first result.}

\item{...}{[\code{ANY}]\cr
Additional arguments passed to to function \code{fun}.}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explictitly passed, uses the last created registry.}
}
\value{
Aggregated results, return type depends on function. If \code{ids} is empty, \code{reduceResults}
  returns \code{init} (if available) or \code{NULL} otherwise.
}
\description{
Reduce Results
}
\seealso{
Other Results: \code{\link{loadResult}};
  \code{\link{reduceResultsDataTable}},
  \code{\link{reduceResultsList}}
}

