% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterFunctions.R
\name{runOSCommand}
\alias{runOSCommand}
\title{Run OS Commands on Local or Remote Machines}
\usage{
runOSCommand(sys.cmd, sys.args = character(0L), nodename = "localhost",
  debug = getOption("batchtools.debug", FALSE))
}
\arguments{
\item{sys.cmd}{[\code{character(1)}]\cr
Command to run.}

\item{sys.args}{[\code{character()}]\cr
Arguments to \code{sys.cmd}.}

\item{nodename}{[\code{character(1)}]\cr
Name of the SSH node to run the command on. If set to \dQuote{localhost} (default), the command
is not piped through SSH.}

\item{debug}{[\code{logical(1)}]\cr
If \code{TRUE}, the commands send to the OS are printed.}
}
\value{
[\code{named list}] with \dQuote{exit.code} (integer) and \dQuote{output} (character).
}
\description{
This is a helper function to run arbitrary OS commands on local or remote machines.
The interface is similar to \code{\link[base]{system2}}, but it always returns the exit status
and the output.
}
\examples{
\dontrun{
runOSCommand("ls")
runOSCommand("ls", "-al")
runOSCommand("notfound")
}
}
\seealso{
Other ClusterFunctions: \code{\link{makeClusterFunctionsDocker}},
  \code{\link{makeClusterFunctionsInteractive}},
  \code{\link{makeClusterFunctionsLSF}},
  \code{\link{makeClusterFunctionsMulticore}},
  \code{\link{makeClusterFunctionsOpenLava}},
  \code{\link{makeClusterFunctionsSGE}},
  \code{\link{makeClusterFunctionsSSH}},
  \code{\link{makeClusterFunctionsSlurm}},
  \code{\link{makeClusterFunctionsSocket}},
  \code{\link{makeClusterFunctionsTorque}},
  \code{\link{makeClusterFunctions}}
}

