% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showLog.r
\name{showLog}
\alias{showLog}
\title{Inspect Log Files}
\usage{
showLog(id, reg = getDefaultRegistry())
}
\arguments{
\item{id}{[\code{integer(1)} or \code{data.table}]\cr
Single integer to specify the job or a \code{data.table} with column \code{job.id}
and exactly one row.}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explicitly passed, uses the last created registry.}
}
\value{
Nothing.
}
\description{
Opens a pager to the stored log file. For customization, see \code{\link[base]{file.show}}.
}
\seealso{
Other debug: \code{\link{killJobs}}, \code{\link{testJob}}
}

