% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/submitJobs.r
\name{submitJobs}
\alias{submitJobs}
\title{Submit jobs or chunks of jobs to batch system via cluster function.}
\usage{
submitJobs(ids = NULL, resources = list(), reg = getDefaultRegistry())
}
\arguments{
\item{ids}{[\code{\link[base]{data.frame}} or \code{integer}]\cr
A \code{\link[base]{data.frame}} (or \code{\link[data.table]{data.table}})
with a column named \dQuote{job.id}.
This is usually the result of a \code{\link{findJobs}} operation, but you may also
pass any other table here, e.g. the return of \code{\link{getJobInfo}}.
Alternatively, you may also pass a vector of integerish job ids.
If not set, defaults to result of \code{\link{findNotSubmitted}}.}

\item{resources}{[\code{list}]\cr
Required resources for all batch jobs. The elements of this list
(e.g. something like \dQuote{walltime} or \dQuote{nodes} depend on by your template job file.
Defaults can be set in the \code{\link{Registry}} via \code{default.resources}.}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explictitly passed, uses the last created registry.}
}
\value{
[\code{\link{data.table}}]. Table with columns \dQuote{job.id} and \dQuote{chunk}.
}
\description{
Submits all jobs to the batch system using the resources provided via \code{resources}.

If an additional column \dQuote{chunk} is present in the table \code{ids},
the jobs will be grouped accordingly. See \code{\link{chunkIds}} for more
information.
}

