% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitJobs.R
\name{submitJobs}
\alias{submitJobs}
\title{Submit Jobs to the Batch Systems}
\usage{
submitJobs(ids = NULL, resources = list(), reg = getDefaultRegistry())
}
\arguments{
\item{ids}{[\code{\link[base]{data.frame}} or \code{integer}]\cr
A \code{\link[base]{data.frame}} (or \code{\link[data.table]{data.table}})
with a column named \dQuote{job.id}.
Alternatively, you may also pass a vector of integerish job ids.
If not set, defaults to the return value of \code{\link{findNotSubmitted}}.}

\item{resources}{[\code{named list}]\cr
  Computational  resources for the batch jobs. The elements of this list
  (e.g. something like \dQuote{walltime} or \dQuote{nodes}) depend on your template file.
  The resources \code{chunk.ncpus} and \code{measure.memory} are reserved for internal functionality:
  The setting \code{chunk.ncpus} is used to determine the number of CPUs to execute jobs in a
  chunk in parallel via \code{mcparallel} (or \code{\link[snow]{sendCall}} on Windows).
  If not set, \code{chunk.ncpus} defaults to 1 (sequential execution).
  The second reserved resource, \code{measure.memory}, can be set to \code{TRUE} to enable the measure of
  memory requirements using \code{\link[base]{gc}}. But note that the reported values are quite
  heuristic and may not reflect the real memory requirements. Furthermore, measuring memory with
  \code{\link[base]{gc}} in parallel is impossible, thus this feature is disabled if \code{chunk.ncpus}
  is greater than one.

  Defaults for all resources, the reserved as well as the those depending on your template, can be stored in the \code{\link{Registry}} in
  \code{default.resources} (e.g., using a configuration file).
  Individual settings set via \code{resources} \code{resources} overrule those in \code{default.resources}.}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explicitly passed, uses the last created registry.}
}
\value{
[\code{\link{data.table}}]. Table with columns \dQuote{job.id} and \dQuote{chunk}.
  See \code{\link{JoinTables}} for examples on working with job tables.
}
\description{
Submits all jobs defined with \code{\link{batchMap}} to the batch system.

If an additional column \dQuote{chunk} is present in the table \code{ids},
the jobs will be grouped accordingly. See \code{\link{chunkIds}} for more
information.

After submitting the jobs, you can use \code{\link{waitForJobs}} to wait for the
termination of jobs or immediately call \code{\link{reduceResultsList}}/\code{\link{reduceResults}}
to collect partial results.
The progress can be monitored with \code{\link{getJobStatus}}.
}

