% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeExperiments.r
\name{summarizeExperiments}
\alias{summarizeExperiments}
\title{Quick Summary over Experiments}
\usage{
summarizeExperiments(ids = NULL, by = c("problem", "algorithm"),
  reg = getDefaultRegistry())
}
\arguments{
\item{ids}{[\code{\link[base]{data.frame}} or \code{integer}]\cr
A \code{\link[base]{data.frame}} (or \code{\link[data.table]{data.table}})
with a column named \dQuote{job.id}.
Alternatively, you may also pass a vector of integerish job ids.
If not set, defaults to all jobs.}

\item{by}{[\code{character}]\cr
Split the resulting table by columns of \code{\link{getJobPars}}.}

\item{reg}{[\code{\link{Registry}}]\cr
Registry. If not explicitly passed, uses the last created registry.}
}
\value{
[\code{named list}] with elements \dQuote{table}
(\code{\link[data.table]{data.table}}] of frequencies), \dQuote{problems}
(\code{character} of problem names) and \dQuote{algorithms}
(\code{character} of algorithm names).
}
\description{
Returns a frequency table of defined experiments.
}

